// src/app/Interfaces/alumnointerface.ts
export class alumnoEntity{    
    idAlumno: number;
    nombre_completo?: string;
    identificacion?: number;
    numero_telefono?:number;
    email?: string; 
    direccion_residencia?: string;    
    genero?:string; 
    numero_matricula?: number; 
    programa?:string; 
    nota_promedio: number;
}